[WIP] start merging PNG prototype with main code
- Split out DataBlobInputStream and DataInputStream to their own files
- Make test of prototype depend on local pngsuite install
- Label PngCodec as an ImageDecoder, have it throw a specific exception
