+++ src/Mike42/GfxPhp/Codec/Common/DataInputStream.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Mike42\GfxPhp\Codec\Common;
+
+interface DataInputStream
+{
+    public function read(int $bytes);
+    public function isEof();
+    public function peek(int $bytes);
+    public function advance(int $bytes);
+}
